Day09: ugly version
This is super ugly.

At least, the first star works on list, but the algorithm is kinda
`O(n)`. However, for the second star, each list is walk for each item to
try for reinsert, so that's definitely `O(n^2)`. It could be refactored
with a different algorithm, keeping the list of free spaces by size AND
by proximity from the beginning. (Althrough the double indexing maybe
costly, I was thinking about as many heap as size, and each heap
contains all the free spaces of size >= reference. So in order to insert
a chunk, just pick the first available space in the relevant heap
(`O(log n)`), but remove them from the other heaps (?) OR store that
this free space was already used (in some set) and clean it later when
`poped` from the other sets.